Added a lot (Look at description)
So I added a quite a bit, I should have been pushing this far more but I forgot. So for one I created a GameGUI class that brings the Player and Table GUI into one window and added code to the Game class that takes the input from the player and add it to the player code. I also added 4 buttons for the gameplay for the user not functional but trying to work on swap. I do have some code for swap but nothing functional and most likely will change it later.